+++ pgsql/pgx/invoiceitem.go
@@ -14,7 +14,7 @@ type InvoiceItem struct {
 }
 
 // Create create item asociated to a header and product for the invoice.
-func (InvoiceItem) Create(tx pgx.Tx, headerID int, items domain.ItemList) error {
+func (InvoiceItem) Create(tx pgx.Tx, headerID int64, items domain.ItemList) error {
 	for _, item := range items {
 		err := tx.QueryRow(context.Background(),
 			`INSERT INTO "invoice_item" (invoice_header_id, product_id) VALUES ($1, $2) RETURNING id, created_at`,
