+++ pgsql/pgx/product_test.go
@@ -61,7 +61,7 @@ func TestProductByID(t *testing.T) {
 
 	tt := []struct {
 		name           string
-		input          int
+		input          int64
 		wantName       string
 		errExpected    bool
 		wantErrContain string
